# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Brave Accounts Service",
        "contact": {}
    },
    "paths": {
...
        "/v2/auth/login/finalize": {
            "post": {
                "description": "Final step of login flow, verifies KE3 message and creates session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Finalize login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + ake token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "login finalize request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginFinalizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginFinalizeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/login/init": {
            "post": {
                "description": "First step of OPAQUE login flow, generates KE2 message.\nEither `blindedMessage`, `clientEphemeralPublicKey` and `clientNonce` must be provided together,\nor `serializedKE1` must be provided.\nIf the latter is provided, `serializedKE2` will be included in the response with other\nKE2 fields omitted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Initialize login",
                "parameters": [
                    {
                        "description": "login init request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginInitRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginInitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/validate": {
            "get": {
                "description": "Validates an auth token and returns session details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Validate auth token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.ValidateTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.LoginFinalizeRequest": {
            "description": "Request to finalize login",
            "type": "object",
            "required": [
                "clientMac"
            ],
            "properties": {
                "clientMac": {
                    "description": "Client MAC of KE3",
                    "type": "string"
                },
                "sessionName": {
                    "description": "Optional name for the new session",
                    "type": "string"
                }
            }
        },
        "controllers.LoginFinalizeResponse": {
            "description": "Response containing auth token after successful login",
            "type": "object",
            "properties": {
                "authToken": {
                    "description": "Authentication token for future requests",
                    "type": "string"
                }
            }
        },
        "controllers.LoginInitRequest": {
            "description": "Request for account login",
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "blindedMessage": {
                    "description": "Blinded message component of KE1",
                    "type": "string"
                },
                "clientEphemeralPublicKey": {
                    "description": "Client ephemeral public key of KE1",
                    "type": "string"
                },
                "clientNonce": {
                    "description": "Client nonce of KE1",
                    "type": "string"
                },
                "email": {
                    "description": "Email address of the account",
                    "type": "string",
                    "example": "test@example.com"
                },
                "serializedKE1": {
                    "description": "Serialized KE1 message",
                    "type": "string"
                }
            }
        },
        "controllers.LoginInitResponse": {
            "description": "Response for account login",
            "type": "object",
            "properties": {
                "akeToken": {
                    "description": "Interim authentication token for future login finalization",
                    "type": "string"
                },
                "evaluatedMessage": {
                    "description": "Evaluated message component of KE2",
                    "type": "string"
                },
                "maskedResponse": {
                    "description": "Server masked response of KE2",
                    "type": "string"
                },
                "maskingNonce": {
                    "description": "Server masking nonce of KE2",
                    "type": "string"
                },
                "serializedKE2": {
                    "description": "Serialized KE2 message",
                    "type": "string"
                },
                "serverEphemeralPublicKey": {
                    "description": "Server ephemeral public key of KE2",
                    "type": "string"
                },
                "serverMac": {
                    "description": "Server MAC of KE2",
                    "type": "string"
                },
                "serverNonce": {
                    "description": "Server nonce of KE2",
                    "type": "string"
                }
            }
        },
...
        "controllers.ValidateTokenResponse": {
            "description": "Response containing validated token details",
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "UUID of the account",
                    "type": "string"
                },
                "email": {
                    "description": "Email address associated with the account",
                    "type": "string"
                },
                "sessionId": {
                    "description": "UUID of the session associated with the account",
                    "type": "string"
                }
            }
        },
...
        "util.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        }
    },
...
}